refactor: move schema startup logic to SchemaManager
